[kernel] Add capabilities to handles
This change finally adds capabilities to handles. Included changes:

- j6_handle_t is now again 64 bits, with the highest 8 bits being a type
  code, and the next highest 24 bits being the capability mask, so that
  programs can check type/caps without calling the kernel.
- The definitions grammar now includes a `capabilities [ ]` section on
  objects, to list what capabilities are relevant.
- j6/caps.h is auto-generated from object capability lists
- init_libj6 again sets __handle_self and __handle_sys, this is a bit
  of a hack.
- A new syscall, j6_handle_list, will return the list of existing
  handles owned by the calling process.
- syscall_verify.cpp.cog now actually checks that the needed
  capabilities exist on handles before allowing the call.

+++ src/kernel/objects/kobject.h
@@ -16,7 +16,7 @@ class kobject
 {
 public:
     /// Types of kernel objects.
-    enum class type : uint16_t
+    enum class type : uint8_t
     {
 #define OBJECT_TYPE( name, val ) name = val,
 #include <j6/tables/object_types.inc>
